requirements (1).txt
streamlit>=1.32.0
streamlit-mermaid>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
validators>=0.22.0